Adding pagination to page Books

+++ src/pages/Books/index.js
@@ -8,30 +8,30 @@ import api from '../../services/api';
 export default function Books() {
 
     const [books, setBooks] = useState([]);
+    const [page, setPage] = useState(1);
 
     const userName = localStorage.getItem('userName');
     const accessToken = localStorage.getItem('accessToken');
+    const authorization = {
+        headers: {
+            Authorization: `Bearer ${accessToken}`
+        }
+    };
     const history = useHistory();
 
     useEffect(() => {
-        api.get('api/Book/v1/asc/25/1', {
-            headers: {
-                Authorization: `Bearer ${accessToken}`
-            }
-        }).then(response => {
-            setBooks(response.data.list);
-        });
-
+        fetchMoreBooks();
     }, [accessToken]);
+
+    async function fetchMoreBooks() {
+        const response = await api.get(`api/Book/v1/asc/4/${page}`, authorization);
+        setBooks([...books, ...response.data.list]);
+        setPage(page + 1);
+    }
 
     async function deleteBook(id) {
         try {
-            await api.delete(`api/Book/v1/${id}`, {
-                headers: {
-                    Authorization: `Bearer ${accessToken}`
-                }
-            });
-
+            await api.delete(`api/Book/v1/${id}`, authorization);
             setBooks(books.filter(book => book.id !== id))
 
         } catch (error) {
@@ -41,11 +41,7 @@ export default function Books() {
 
     async function logout() {
         try {
-            await api.get('api/auth/v1/revoke', {
-                headers: {
-                    Authorization: `Bearer ${accessToken}`
-                }
-            });
+            await api.get('api/auth/v1/revoke', authorization);
 
             history.push('/');
             localStorage.clear();
@@ -99,6 +95,9 @@ export default function Books() {
                     );
                 })}
             </ul>
+            <button className="button" type="button" onClick={fetchMoreBooks}>
+                Load more
+            </button>
         </div>
     )
 }
